+++ lib/workers/repository.js
@@ -1,6 +1,7 @@
 // Global requires
 const handlebars = require('handlebars');
 const ini = require('ini');
+let logger = require('../helpers/logger');
 const stringify = require('json-stringify-pretty-compact');
 // API
 const githubApi = require('../api/github');
@@ -20,7 +21,6 @@ module.exports = {
 
 // This will be github or others
 let api;
-let logger;
 
 // Queue package files in sequence within a repo
 async function processRepo(config) {
