+++ test/workers/branch.spec.js
@@ -1,4 +1,5 @@
 const branchWorker = require('../../lib/workers/branch');
+const prWorker = require('../../lib/workers/pr');
 const npmHelper = require('../../lib/helpers/npm');
 const yarnHelper = require('../../lib/helpers/yarn');
 const defaultConfig = require('../../lib/config/defaults').getConfig();
@@ -302,6 +303,50 @@ describe('workers/branch', () => {
       expect(npmHelper.getLockFile.mock.calls.length).toBe(0);
       expect(yarnHelper.maintainLockFile.mock.calls.length).toBe(1);
       expect(config.api.commitFilesToBranch.mock.calls.length).toBe(0);
+    });
+  });
+  describe('updateBranch(upgrades)', () => {
+    let config;
+    beforeEach(() => {
+      config = Object.assign({}, defaultConfig);
+      config.api = {
+        checkForClosedPr: jest.fn(),
+      };
+      branchWorker.ensureBranch = jest.fn();
+      prWorker.ensurePr = jest.fn();
+    });
+    it('returns immediately if closed PR found', async () => {
+      config.api.checkForClosedPr.mockReturnValue(true);
+      await branchWorker.updateBranch([config]);
+      expect(branchWorker.ensureBranch.mock.calls.length).toBe(0);
+    });
+    it('does not return immediately if recreateClosed true', async () => {
+      config.api.checkForClosedPr.mockReturnValue(true);
+      config.recreateClosed = true;
+      await branchWorker.updateBranch([config]);
+      expect(branchWorker.ensureBranch.mock.calls.length).toBe(1);
+    });
+    it('pins', async () => {
+      config.upgradeType = 'pin';
+      await branchWorker.updateBranch([config]);
+      expect(branchWorker.ensureBranch.mock.calls.length).toBe(1);
+    });
+    it('majors', async () => {
+      config.upgradeType = 'major';
+      await branchWorker.updateBranch([config]);
+      expect(branchWorker.ensureBranch.mock.calls.length).toBe(1);
+    });
+    it('minors', async () => {
+      config.upgradeType = 'minor';
+      await branchWorker.updateBranch([config]);
+      expect(branchWorker.ensureBranch.mock.calls.length).toBe(1);
+    });
+    it('handles errors', async () => {
+      config.api.checkForClosedPr = jest.fn(() => {
+        throw new Error('oops');
+      });
+      await branchWorker.updateBranch([config]);
+      expect(branchWorker.ensureBranch.mock.calls.length).toBe(0);
     });
   });
 });
